+++ rust/cardano-blockchain-types/src/lib.rs
@@ -1,10 +1,10 @@
 //! Catalyst Enhanced `MultiEraBlock` Structures
 
 mod auxdata;
-mod cip36;
 pub mod conversion;
 mod fork;
 pub mod hashes;
+mod metadata;
 mod multi_era_block_data;
 mod network;
 mod point;
@@ -21,11 +21,11 @@ pub use auxdata::{
     metadatum_value::MetadatumValue,
     scripts::{Script, ScriptArray, ScriptType, TransactionScripts},
 };
-pub use cip36::{
+pub use fork::Fork;
+pub use metadata::cip36::{
     key_registration::Cip36KeyRegistration, registration_witness::Cip36RegistrationWitness,
     voting_pk::VotingPubKey, Cip36, Cip36Validation,
 };
-pub use fork::Fork;
 pub use multi_era_block_data::MultiEraBlock;
 pub use network::Network;
 pub use point::Point;
