+++ rust/cardano-blockchain-types/src/metadata/cip36/registration_witness.rs
