+++ rust/cardano-blockchain-types/src/metadata/mod.rs
@@ -1,0 +1,3 @@
+//! Metadata module
+
+pub mod cip36;
